Fix for tab bar
