+++ ts/navigation/TabNavigator.tsx
@@ -1,8 +1,7 @@
 import { IOColors } from "@pagopa/io-app-design-system";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import React from "react";
-import { StyleSheet } from "react-native";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { useBottomTabNavigatorStyle } from "../hooks/useBottomTabNavigatorStyle";
 import LoadingSpinnerOverlay from "../components/LoadingSpinnerOverlay";
 import { makeFontStyleObject } from "../components/core/fonts";
 import { TabIconComponent } from "../components/ui/TabIconComponent";
@@ -20,7 +19,6 @@ import {
   isSettingsVisibleAndHideProfileSelector
 } from "../store/reducers/backendStatus";
 import { StartupStatusEnum, isStartupLoaded } from "../store/reducers/startup";
-import variables from "../theme/variables";
 import { MESSAGES_ROUTES } from "../features/messages/navigation/routes";
 import { SERVICES_ROUTES } from "../features/services/common/navigation/routes";
 import { showBarcodeScanSection } from "../config";
@@ -31,30 +29,8 @@ import ROUTES from "./routes";
 
 const Tab = createBottomTabNavigator<MainTabParamsList>();
 
-export const MyStyles = StyleSheet.create({
-  tabBarStyle: {
-    backgroundColor: IOColors.white,
-    paddingLeft: 3,
-    paddingRight: 3,
-    borderTopWidth: 0,
-    paddingTop: 8,
-    // iOS shadow
-    shadowColor: variables.footerShadowColor,
-    shadowOffset: {
-      width: variables.footerShadowOffsetWidth,
-      height: variables.footerShadowOffsetHeight
-    },
-    zIndex: 999,
-    shadowOpacity: variables.footerShadowOpacity,
-    shadowRadius: variables.footerShadowRadius,
-    // Android shadow
-    elevation: variables.footerElevation
-  }
-});
-
 export const MainTabNavigator = () => {
   const navigation = useIONavigation();
-  const insets = useSafeAreaInsets();
 
   const startupLoaded = useIOSelector(isStartupLoaded);
   const isDesignSystemEnabled = useIOSelector(isDesignSystemEnabledSelector);
@@ -66,13 +42,11 @@ export const MainTabNavigator = () => {
     isSettingsVisibleAndHideProfileSelector
   );
 
-  const tabBarHeight = 54;
-  const additionalPadding = 10;
-  const bottomInset = insets.bottom === 0 ? additionalPadding : insets.bottom;
-
   const navigateToBarcodeScanScreen = () => {
     navigation.navigate(ROUTES.BARCODE_SCAN);
   };
+
+  const tabBarStyle = useBottomTabNavigatorStyle();
 
   return (
     <LoadingSpinnerOverlay
@@ -100,11 +74,7 @@ export const MainTabNavigator = () => {
             ? IOColors["blueIO-500"]
             : IOColors.blue,
           tabBarInactiveTintColor: IOColors["grey-850"],
-          tabBarStyle: [
-            MyStyles.tabBarStyle,
-            { height: tabBarHeight + bottomInset },
-            insets.bottom === 0 ? { paddingBottom: additionalPadding } : {}
-          ]
+          tabBarStyle
         }}
       >
         <Tab.Screen
